+++ sdk/storage/azure-storage-files-datalake/src/datalake_file_system_client.cpp
@@ -124,7 +124,9 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     policies.emplace_back(std::make_unique<Storage::Details::StoragePerRetryPolicy>());
 
     {
-      Azure::Core::Http::GetTokenOptions const tokenOptions = {{Storage::Details::StorageScope}};
+      Azure::Core::Http::TokenRequestOptions const tokenOptions
+          = {{Storage::Details::StorageScope}};
+
       policies.emplace_back(std::make_unique<Azure::Core::Http::BearerTokenAuthenticationPolicy>(
           credential, tokenOptions));
     }